pyproject.toml
[tool.poetry]
name = "tvae"
version = "0.1.0"
description = ""
authors = ["PhilippeRacicot <10731763+Vaunorage@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "< 1.23.0"
fastai = "^2.7.10"
optuna = "^3.1.0"
matplotlib = "^3.6.3"
tqdm = "^4.64.1"
torchmetrics = "^0.11.1"
jupyter = "^1.0.0"
seaborn = "*"
plotly = "^5.13.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
